Don't attach helm resources to steps
Cleanup is based on removing resources associated with steps, in reverse
order. But, there is no need to cleanup helm-created resources, as they
will all be destroyed with the cluster nodegroup is cleaned up, so we
can avoid wasting time cleaning them up, or giving the user unnecessary
instructions at cleanup.